refact(add-party): extracted a party-type-agnostic replication core
Moving a party onto a participant that does not hold it yet is the same
sequence of Canton calls whoever owns the party: capture an offset before the
topology moves, export the ACS scoped to the target, import it across a
synchronizer disconnect, clear the Onboarding marker. None of it needs a
DecentralizedNamespaceDefinition, but all of it took an AddPartyConfig, which
tied it to decparties for no reason other than where it happened to live.

The core now takes a ReplicationTarget — a party, a participant, and the
artifact keys one run's durable markers live under. add_party supplies its own
keys via ADD_PARTY_REPLICATION_ARTIFACTS, whose values are the strings it has
always used, so a run interrupted before this still resumes after it.

What stayed behind is what is genuinely party-type specific: deciding the
topology and getting it authorized. A decparty needs owner-threshold
signatures over a DNS; that lives with the workflow, not the core.

No behavior change. Only artifact-missing error text differs, and it now names
the key it looked for.
