+++ crates/decman/src/workflow/add_party/config.rs
@@ -2,7 +2,23 @@ use std::marker::PhantomData;
 
 use serde::{Deserialize, Serialize};
 
-use crate::canton_id::CantonId;
+use crate::{
+    canton_id::CantonId,
+    workflow::{
+        party_replication::{ReplicationArtifacts, ReplicationTarget},
+        storage::artifact_kinds,
+    },
+};
+
+/// The artifact keys the add-party run's replication has always used. Named
+/// here rather than inside the replication core so the core stays free of any
+/// one workflow's key namespace — and so these strings keep their exact values,
+/// which is what lets a run interrupted before the extraction resume after it.
+pub const ADD_PARTY_REPLICATION_ARTIFACTS: ReplicationArtifacts = ReplicationArtifacts {
+    export_offset: artifact_kinds::ADD_PARTY_EXPORT_OFFSET,
+    pre_activation_offset: artifact_kinds::ADD_PARTY_PRE_ACTIVATION_OFFSET,
+    import_inflight: artifact_kinds::ADD_PARTY_ACS_IMPORT_INFLIGHT,
+};
 
 /// Configuration for the add-party workflow (adding a new member to an
 /// existing decentralized party)
@@ -56,5 +72,19 @@ impl AddPartyConfig {
     /// Daml signing key name for the new member (see `namespace_key_name`).
     pub fn daml_key_name(&self) -> String {
         format!("{}-daml-transactions", self.decentralized_party_id.prefix)
+    }
+
+    /// This run's replication, addressed the way the party-type-agnostic core
+    /// expects: the decentralized party moving onto the new member.
+    ///
+    /// `instance_name` is passed rather than read off `self` because the step
+    /// callers already carry it, and the two must not be allowed to disagree.
+    pub fn replication_target(&self, instance_name: &str) -> ReplicationTarget {
+        ReplicationTarget::new(
+            self.decentralized_party_id.clone(),
+            self.new_participant_id.clone(),
+            instance_name.to_string(),
+            ADD_PARTY_REPLICATION_ARTIFACTS,
+        )
     }
 }
