+++ crates/decman/src/workflow/add_party/steps/mod.rs
@@ -1,14 +1,9 @@
-pub mod acs_sync;
 pub mod clear_onboarding;
 pub mod export_state;
 pub mod generate_keys;
 pub mod proposals;
 
-pub use acs_sync::{collect_party_package_ids, export_party_acs, import_party_acs};
-pub use clear_onboarding::{
-    ClearOutcome, author_clear_proposal, clear_onboarding_flag, sign_clear_proposal,
-    submit_clear_proposal,
-};
+pub use clear_onboarding::{author_clear_proposal, sign_clear_proposal, submit_clear_proposal};
 pub use export_state::export_state;
 pub use generate_keys::generate_keys;
 pub use proposals::{create_proposals, sign_proposals, submit_proposals};
